moduller/requirements.txt
earthengine-api
folium
branca
requests
Pillow
matplotlib
pandas
numpy
jupyter
geopandas
